requirements.txt
fastmcp>=0.4.1
httpx>=0.26.0
pydantic>=2.5.3
python-dotenv>=1.0.0
